requirements.txt
requests==2.32.3
parsel==1.10.0
pycryptodome==3.22.0